requirements.txt
streamlit
opencv-python
torch
Pillow
ultralytics
scikit-learn
